Fixed the navbar, so authentication is included and everything is working for authentication and where it should be in the navbar.
Next: At this point, I need to connect the data I am getting from the
back end with the front end so I have data to display. If I can get this
figured out by the end of the day, I can focus on caching the back end
data to get to exceeds expectations, but I need to get the data rendered
properly for now, which is a tall order.